evaluate complete, userpage img

+++ src/main/java/com/court/supporter/command/TB_013VO.java
@@ -13,12 +13,12 @@ import lombok.NoArgsConstructor;
 public class TB_013VO {
 
 	private String evaluate_proper_num;
-	private Integer all_career_score;
-	private Integer jrsdc_career_score;
-	private Integer office_score;
-	private Integer certificate_score;
-	private Integer adjust_score;
-	private Integer evaluate_score;
+	private Double all_career_score;
+	private Double jrsdc_career_score;
+	private Double office_score;
+	private Double certificate_score;
+	private Double adjust_score;
+	private Double evaluate_score;
 	private String review_etc;
 	private String admin_proper_num; 
 	private String aplcn_dtls_proper_num;
